gitcmd: graphql resolver fixes for grep
* LineMatch needs to be a pointer for graphql
* Use int for lineNumber

+++ api/schema.go
@@ -148,7 +148,7 @@ type FileMatch {
 
 type LineMatch {
 	preview: String!
-	lineNumber: Float!
+	lineNumber: Int!
 }
 
 type DependencyReferences {
